+++ packages/@headlessui-react/src/components/listbox/listbox.tsx
@@ -537,9 +537,9 @@ function ListboxFn<
     if (defaultValue === undefined) return
 
     d.addEventListener(form.current, 'reset', () => {
-      onChange(defaultValue)
+      theirOnChange?.(defaultValue)
     })
-  }, [form, onChange /* Explicitly ignoring `defaultValue` */])
+  }, [form, theirOnChange /* Explicitly ignoring `defaultValue` */])
 
   return (
     <ListboxActionsContext.Provider value={actions}>
@@ -680,7 +680,7 @@ function ButtonFn<TTag extends ElementType = typeof DEFAULT_BUTTON_TAG>(
     type: useResolveButtonType(props, data.buttonRef),
     'aria-haspopup': 'listbox',
     'aria-controls': data.optionsRef.current?.id,
-    'aria-expanded': data.disabled ? undefined : data.listboxState === ListboxStates.Open,
+    'aria-expanded': data.listboxState === ListboxStates.Open,
     'aria-labelledby': labelledby,
     disabled: data.disabled,
     onKeyDown: handleKeyDown,