+++ packages/@headlessui-tailwindcss/CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Nothing yet!
 
+## [0.2.0] - 2023-07-27
+
+### Added
+
+- Add `ui-focus-visible` variant to Tailwind plugin ([#2347](https://github.com/tailwindlabs/headlessui/pull/2347))
+- Add `ui-not-focus-visible` variant to Tailwind plugin ([#2618](https://github.com/tailwindlabs/headlessui/pull/2618))
+
 ## [0.1.3] - 2023-04-12
 
 ### Fixed
@@ -33,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `@headlessui/tailwindcss` plugin ([#1487](https://github.com/tailwindlabs/headlessui/pull/1487))
 
-[unreleased]: https://github.com/tailwindlabs/headlessui/compare/@headlessui/tailwindcss@v0.1.3...HEAD
+[unreleased]: https://github.com/tailwindlabs/headlessui/compare/@headlessui/tailwindcss@v0.2.0...HEAD
+[0.2.0]: https://github.com/tailwindlabs/headlessui/compare/@headlessui/tailwindcss@v0.1.3...@headlessui/tailwindcss@v0.2.0
 [0.1.3]: https://github.com/tailwindlabs/headlessui/compare/@headlessui/tailwindcss@v0.1.2...@headlessui/tailwindcss@v0.1.3
 [0.1.2]: https://github.com/tailwindlabs/headlessui/compare/@headlessui/tailwindcss@v0.1.1...@headlessui/tailwindcss@v0.1.2
 [0.1.1]: https://github.com/tailwindlabs/headlessui/compare/@headlessui/tailwindcss@v0.1.0...@headlessui/tailwindcss@v0.1.1